enable JsonIdentityInfoAndBackReferencesTest
